No more DISPLAY_WIDTH and DISPLAY_HEIGHT OK + struct vgaboard_t

+++ include/hagl_hal_color.h
@@ -31,6 +31,8 @@ SPDX-License-Identifier: MIT
 
 */
 
+#include <stdint.h>
+
 #ifndef _HAGL_GD_HAL_COLOR_H
 #define _HAGL_GD_HAL_COLOR_H
 
@@ -38,12 +40,19 @@ SPDX-License-Identifier: MIT
 extern "C" {
 #endif
 
-#include <stdint.h>
-
-/** HAL must provide typedef for colors. This HAL uses RGB565. */
-typedef uint32_t color_t;
+/** 
+ * @brief HAL must provide typedef for colors.
+ * This HAL uses RGB555 in 15bpp "true color" mode, 
+ * or an index for paletted modes at 
+ *  => 1bpp (0-1, mask 0x1), 
+ *  => 2bpp (0-3, mask 0x3), 
+ *  => 4bpp (0-15, mask 0xf) and 
+ *  => 8bpp (0-255, mask 0xff).
+ */
+typedef uint16_t color_t;
 
 #ifdef __cplusplus
 }
 #endif
+
 #endif /* _HAGL_GD_HAL_COLOR_H */
